feat: load extension list from an external file
